# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.ScenarioRequest:
    properties:
      aero_coeff:
        type: number
      description:
        type: string
      name:
        type: string
      rolling_coeff:
        type: number
      speed:
        type: number
      system_consumption:
        type: number
      type:
        enum:
        - дорога
        - комфорт
        type: string
    required:
    - name
    - type
    type: object
  handler.ScenarioResponse:
    properties:
      aero_coeff:
        type: number
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      rolling_coeff:
        type: number
      speed:
        type: number
      status:
        type: string
      system_consumption:
        type: number
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tesla-app.com
    name: API Support
    url: http://localhost:8080
  description: API for Tesla driving scenarios and trip management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Tesla App API
  version: "1.0"
paths:
  /api/scenarios:
    get:
      consumes:
      - application/json
      description: Get all driving scenarios with optional filtering
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ScenarioResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Get scenarios list
      tags:
      - Scenarios
    post:
      consumes:
      - application/json
      description: Create new driving scenario (moderator only)
      parameters:
      - description: Scenario data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.ScenarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.ScenarioResponse'
        "400":
          description: Bad request
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create scenario
      tags:
      - Scenarios
  /api/scenarios/{id}:
    delete:
      consumes:
      - application/json
      description: Delete driving scenario (moderator only)
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            type: object
        "404":
          description: Scenario not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete scenario
      tags:
      - Scenarios
    get:
      consumes:
      - application/json
      description: Get driving scenario details by ID
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ScenarioResponse'
        "400":
          description: Invalid ID
          schema:
            type: object
        "404":
          description: Scenario not found
          schema:
            type: object
      summary: Get scenario by ID
      tags:
      - Scenarios
    put:
      consumes:
      - application/json
      description: Update driving scenario (moderator only)
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      - description: Scenario data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.ScenarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ScenarioResponse'
        "400":
          description: Bad request
          schema:
            type: object
        "404":
          description: Scenario not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update scenario
      tags:
      - Scenarios
  /api/scenarios/{id}/add-to-trip:
    post:
      consumes:
      - application/json
      description: Add driving scenario to user's draft trip
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      - description: Duration data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Scenario added to trip
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add scenario to trip
      tags:
      - Scenarios
  /api/scenarios/{id}/scenarioimage:
    post:
      consumes:
      - multipart/form-data
      description: Upload image for driving scenario (moderator only)
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Upload scenario image
      tags:
      - Scenarios
...
  /api/trips/{trip_id}/scenarios/{scenario_id}:
    delete:
      consumes:
      - application/json
      description: Remove scenario from trip (only for draft trips by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Scenario ID
        in: path
        name: scenario_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Scenario removed from trip
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove scenario from trip
      tags:
      - TripScenarios
    put:
      consumes:
      - application/json
      description: Update scenario duration in trip (only for draft trips by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Scenario ID
        in: path
        name: scenario_id
        required: true
        type: integer
      - description: Duration data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Trip scenario updated
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update trip scenario
      tags:
      - TripScenarios
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"